// src/shared/types.ts
interface IconAvatar {
  type: "icon"
  icon: string
  hue?: number
  title?: string
}
// ...
interface ApiResultItem {
  type: "text" | "file" | "quote"
  contentText?: string
  contentBuffer?: ArrayBuffer
  name?: string
  mimeType?: string
}